Add operations-utilities dependency, switch to operations-utils evm utils
